Added the currently available Films to home screen These link to the Club rep showings for that Films
Note: Since Docker is still not working for me,
all these changes were made on a seperate django project
and then ported over to this repo. variable names, as well as models,
are all different. And since Docker doesn't work, I cannot test this
So if something is broken that is why. Msg me on discord if that is the case
- Owain